requirements.txt
# django==1.11.7
# django-filter==1.1.0
# django-formtools==2.1
# django-material==1.1.1
# django-viewflow==1.1.0

Django==4.1.2
django-filter==22.1
django-formtools==2.4
django-material==1.11.3
django-viewflow==1.11.0

# six==1.16.0